requirements.txt
strands-agents>=1.15.0
fastapi>=0.104.0
streamlit>=1.28.0
pillow>=10.0.0
openai>=1.3.0
python-dotenv>=1.0.0
mcp>=0.1.0
python-multipart>=0.0.6
requests>=2.31.0
matplotlib>=3.7.0
geopy>=2.4.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
ruff>=0.1.0
pre-commit>=3.5.0
mypy>=1.7.0
types-requests>=2.31.0

